requirements.txt
streamlit
openai
langchain
langchain-openai
langchain-community
faiss-cpu
tiktoken
python-dotenv
